+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pick-All strategy
 - OLS linear regression calculation 
 ### Fixed
+
 ### Changed
+- allow quandl pull without ticker list
+
 ### Removed