+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pick-All strategy
 - OLS linear regression calculation 
 - test folder for mathstats
+- overwrite history function for Snowflake
 ### Fixed
 
 ### Changed