+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Pick-All strategy
 - OLS linear regression calculation 
+- ridge regression calculation 
 - test folder for mathstats
 - overwrite history function for Snowflake
 ### Fixed